+++ src/reducers/index.js
@@ -1,5 +1,6 @@
 import account from './_account';
 import send from './_send';
+import transactions from './_transactions';
 import {
   accountChangeLanguage,
   accountChangeNativeCurrency,
@@ -12,6 +13,12 @@ import {
   accountUpdateTransactions,
   INITIAL_ACCOUNT_STATE,
 } from './_account';
+import {
+  transactionsClearState,
+  transactionsGetAccountTransactions,
+  transactionsUpdateHasPendingTransaction,
+  transactionsUpdateTransactions,
+} from './_transactions';
 import {
   sendClearFields,
   sendMaxBalance,
@@ -32,9 +39,7 @@ export {
   accountInitializeState,
   accountUpdateAccountAddress,
   accountUpdateExchange,
-  accountUpdateHasPendingTransaction,
   accountUpdateNetwork,
-  accountUpdateTransactions,
   INITIAL_ACCOUNT_STATE,
   send,
   sendClearFields,
@@ -47,4 +52,9 @@ export {
   sendUpdateNativeAmount,
   sendUpdateRecipient,
   sendUpdateSelected,
+  transactions,
+  transactionsClearState,
+  transactionsGetAccountTransactions,
+  transactionsUpdateHasPendingTransaction,
+  transactionsUpdateTransactions,
 };